step11/gae-project/requirements.txt
Flask
google-cloud-datastore
google-cloud-storage
gunicorn
google-cloud-core
